Fixed a number of warnings from the static analyzer and switched to GetTickCount64 (or a surrogate function on pre-Vista) whereever we use millisecond values

+++ windirstat/Controls/pacman.h
@@ -49,7 +49,7 @@ private:
     double m_position;      // 0...1
     bool m_mouthOpening;    // Mouth is opening
     double m_aperture;      // 0...1
-    DWORD m_lastUpdate;     // TickCount
+    ULONGLONG m_lastUpdate; // TickCount
 };
 
 #endif // __WDS_PACMAN_H__
